requirements.txt
python-dotenv
rouge-score
numpy
tqdm
openai==0.28.1